requirements.txt
psycopg2
azure-identity